requirements-pi.txt
RPi.GPIO
mfrc522
requests
python-dotenv
